+++ src/http/request.rs
@@ -1,0 +1,7 @@
+use super::Method;
+
+pub struct Request {
+    path: String,
+    query_string: Option<String>,
+    method: Method,
+}
